log which command a user supplied if it is unknown
DISCLAIMER: the code here does what's like a switch case over hard-coded values of commands, and when it finds them, it does print them, but when it was not found, it wasn't printed. I'm not sure about the security implications of adding $command here, but I couldn't find any information about exploits in the perl join or print statements, so I'm using it anyway.  Use at your own risk.